competitions/requirements.txt
numpy==1.21.6
pandas==1.3.5
scipy==1.7.2
matplotlib==3.5.0
scikit-learn==1.0.1
statsmodels==0.12.1
quantstats==0.0.59
openpyxl==3.0.10
gcastle==1.0.2
tigramite==5.0.0.3
ylearn==0.1.3
cdt==0.5.23
graphviz==0.20.1
jupyter==1.0.0